move extended aliases build steps from the server to scripts folder
Preparing the idea to have extended type aliases as a built asset, rather than a JIT object with cached intermediary steps.
The main avantage to this approach would be that this long-to-build asset does not need to be rebuilt from scratch when running test from empty databases
