+++ scripts/entities_extended_types_aliases/build_extended_aliases.ts
@@ -1,19 +1,20 @@
+import { writeFile } from 'node:fs/promises'
 import { difference, intersection, uniq, values, omit, cloneDeep } from 'lodash-es'
 import { prefixifyWd } from '#controllers/entities/lib/prefix'
 import { makeCachedSparqlRequest } from '#data/wikidata/make_sparql_request'
-import { extendedAliasesQueries } from '#data/wikidata/queries/extended_type_aliases_queries'
+import { absolutePath } from '#lib/absolute_path'
 import { cache_ } from '#lib/cache'
 import { newError } from '#lib/error/error'
 import { oneDay, oneYear } from '#lib/time'
 import { getHashCode, objectEntries } from '#lib/utils/base'
 import { info, logError } from '#lib/utils/logs'
 import { getTypesFromTypesAliases, typesAliases, type PluralizedEntityType } from '#lib/wikidata/aliases'
+import { extendedAliasesQueries } from '#scripts/entities_extended_types_aliases/extended_type_aliases_queries'
 import type { WdEntityId, WdEntityUri } from '#server/types/entity'
 
 export const extendedTypesAliases = cloneDeep(typesAliases)
 const stats = {}
 
-// Let scripts/refresh_entities_type_extended_aliases.sh force a refresh by setting an environment variable
 const refresh = process.env.INV_REFRESH_ENTITIES_TYPE_EXTENDED_ALIASES === 'true'
 
 for (const [ type, sparql ] of objectEntries(extendedAliasesQueries)) {
@@ -109,3 +110,7 @@ function dropOverlaps (type: PluralizedEntityType, typeExtendedAliases: WdEntity
 }
 
 export const typesByExtendedP31AliasesValues = getTypesFromTypesAliases(extendedTypesAliases)
+
+const path = absolutePath('server', 'assets/extended_types_aliases.json')
+await writeFile(path, JSON.stringify(extendedTypesAliases, null, 2))
+info(`Extended entities types aliases saved in ${path}`)
