+++ scripts/entities_extended_types_aliases/extended_type_aliases_queries.ts
